+++ examples/basic_window.rb
@@ -5,7 +5,7 @@ require 'glimmer-dsl-libui'
 include Glimmer
 
 window('hello world', 300, 200, 1) {
-#   on_closing do
-#     puts 'Bye Bye'
-#   end
+  on_closing do
+    puts 'Bye Bye'
+  end
 }.show