+++ lib/glimmer/dsl/libui/dsl.rb
@@ -40,6 +40,7 @@ module Glimmer
 #           block_attribute
 #           attribute
         %w[
+          listener
           control
         ]
       )