+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ## 0.0.1
 
+- LibUI general control and window support
+- LibUI listener support
 - Support examples/basic_window.rb